Fraud-Detection-in-UPI---FDT2/requirements.txt
APScheduler==3.11.2
PyJWT==2.10.1
PyYAML==6.0.3
aioredis==2.0.1
bcrypt==5.0.0
fastapi==0.128.0
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.5.3
numpy==2.4.2
pandas==3.0.0
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.4
python-multipart==0.0.6
redis==7.1.0
scikit-learn==1.8.0
starlette==0.50.0
uvicorn==0.40.0
xgboost==3.1.3
groq>=0.9.0
